+++ CollisionUtil.hh
@@ -7,17 +7,17 @@
 namespace collision{
 
 //There's probably a much simpler way to do all this ...
-static bool pointInsideBox(point_t point, point_t boxCenter, double boxLenX, double boxLenY)
+static bool pointInsideBox(point_t point, point_t boxCenter, float boxLenX, float boxLenY)
 {
-    return abs(point.x - boxCenter.x) < boxLenX / 2.0 && abs(point.y - boxCenter.y) < boxLenY / 2.0;
+    return abs(point.x - boxCenter.x) < boxLenX / 2.0f && abs(point.y - boxCenter.y) < boxLenY / 2.0f;
 }
 
-static bool pointInsideCircle(point_t point, point_t circleCenter, double circleRadius)
+static bool pointInsideCircle(point_t point, point_t circleCenter, float circleRadius)
 {
     return math_util::dist(point, circleCenter) < circleRadius;
 }
 
-static bool verticalLineCircle(point_t bottom, point_t top, point_t circleCenter, double circleRadius)
+static bool verticalLineCircle(point_t bottom, point_t top, point_t circleCenter, float circleRadius)
 {
     return abs(circleCenter.x - bottom.x) < circleRadius && (
         (circleCenter.y > bottom.y && circleCenter.y < top.y) ||
@@ -25,7 +25,7 @@ static bool verticalLineCircle(point_t bottom, point_t top, point_t circleCenter
         pointInsideCircle(top, circleCenter, circleRadius));
 }
 
-static bool horizontalLineCircle(point_t left, point_t right, point_t circleCenter, double circleRadius)
+static bool horizontalLineCircle(point_t left, point_t right, point_t circleCenter, float circleRadius)
 {
     return abs(circleCenter.y - left.y) < circleRadius && (
         (circleCenter.x > left.x && circleCenter.x < right.x) ||
@@ -33,7 +33,7 @@ static bool horizontalLineCircle(point_t left, point_t right, point_t circleCent
         pointInsideCircle(right, circleCenter, circleRadius));
 }
 
-static bool boxCircle(point_t boxCenter, point_t boxSize, point_t circleCenter, double circleRadius)
+static bool boxCircle(point_t boxCenter, point_t boxSize, point_t circleCenter, float circleRadius)
 {
     if(pointInsideBox(circleCenter, boxCenter, boxSize.x, boxSize.y))
     {
@@ -45,11 +45,11 @@ static bool boxCircle(point_t boxCenter, point_t boxSize, point_t circleCenter, 
     }
     else
     {
-        point_t halfBox = boxSize / 2.0;
+        point_t halfBox = boxSize / 2.0f;
         point_t topLeft = boxCenter + point_t(-1.0 * halfBox.x, halfBox.y);
         point_t topRight = boxCenter + halfBox;
         point_t bottomLeft = boxCenter - halfBox;
-        point_t bottomRight = boxCenter + point_t(halfBox.x, -1.0 * halfBox.y);
+        point_t bottomRight = boxCenter + point_t(halfBox.x, -1.0f * halfBox.y);
 
         return horizontalLineCircle(topLeft, topRight, circleCenter, circleRadius) ||
                horizontalLineCircle(bottomLeft, bottomRight, circleCenter, circleRadius) ||
